Correcciones en clases
Cambio de Double por BigDecimal en carrito.
Modificación por @NotBlank en client, admin y developer

+++ src/main/java/ExperienceGroup/Ludora/features/cart/domain/dto/CartDTOResponse.java
@@ -4,10 +4,11 @@ package ExperienceGroup.Ludora.features.cart.domain.dto;
 import ExperienceGroup.Ludora.features.client.domain.dto.ClientDTOResponse;
 import ExperienceGroup.Ludora.features.game.domain.dto.GameDTOResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartDTOResponse(
         ClientDTOResponse client,
         List<GameDTOResponse> games,
-        Double totalPrice
+        BigDecimal totalPrice
 ) {}
